feat(Devnode.NodeList): create a store abstraction for nodes
The `Devnode.NodeList` module provides an abstraction for storing nodes.
It is built by composing the `HashDict`.

The list supports the following operations:
 * Creating a new `NodeList`
 * Adding new entries
 * Getting an entry

A `NodeList` entry is a Map consisting of a unique `name` and an `ip`.
The `name` is user defined, but the ip is internally assigned to the
entry by the `NodeList`.