feat(Devnode.Monitor.NodeServer): Server for managing a NodeList
The `Devnode.Monitor.NodeServer` module is a `GenServer` for managing a
`NodeList`. The module provides a client API with the following
functions:

 * start_link: starts a globally registered GenServer
 * add_entry: adds a new node to the `NodeList`
 * entries: lists nodes in the `NodeList`
 * purge: clears the entries in the `NodeList`

The `Devnode.Monitor` `start` function has been modified to start the
`Devnode.Monitor.NodeServerSupervisor` which starts the
`Devnode.Monitor.NodeServer` as a worker with a `one_for_one` restart
strategy.