feat(ImageRepo): add basic ImageRepo and FakeImageRepo
The purpose of the `Devnode.Client.ImageRepo` module is to provide
functionality relating to repository of Dockerfiles available to
the Devnode CLI.

Currently the `Devnode.Client.ImageRepo` only contains one function
`dir` which returns a hard-coded file path to the location of the
repository. In the future this path will be configurable.

A `FakeImageRepo` module has been provided for testing purposes. This
module provides the ability to build a fake image repository, that can
be used as a test double.