feat(Monitor.NodeServer): maintain state when server is restarted
The `Monitor.NodeServer` module now persists server state when a new
entry is added, or the server state is purged.

A new module called `NodeStash` is reponsible for converting the state
of the server (a `Monitor.NodeList`) to binary and writing to a file.
The location of this file is configured with the aid of the `Mix.config`
module. The `devnode_monitor` app has been setup to load different
configuration in the test environment, which enables a different `stash`
file path to be used for the tests.

When `Monitor.NodeServer` is initialised it attempts to read the data
from the file, reinstating the previous server state.