feat(Client.CLI): execute command in a task and monitor
The `Devnode.Client.Command.execute` function is now executed within a
linked process. This is done so that if the linked process crashes
the parent process can trap the exits and respond gracefully.

The `Client.CLI` handles these potential crashes by generating a `help`
response for the relevant command.

The `Devnode.Client.Command.build_node/1` function has been updated to
internally call `scaffold_node/1` with the required `name` option. The
`scaffold_node/1` function implements a guard that checks that the
supplied `name` argument is a non-empty string. This ensures that if the
`name` option is not supplied then a `FunctionClauseError` is raised.
This approach of "programming to the happy path" is facilitated by the
fact that the parent process is trapping errors.