refactor(ImageRepo): read image_repo path from config
Both the `Client.ImageRepo` and `Support.FakeImageRepo` modules now
read the path for image repository directory from config.

In the `devnode_client` application the `image_repo` path is overridden
in the test environment config. This has simplified the client tests as
it removes the need to mock the `Devnode.Client.ImageRepo.dir` function.