feat(RuntimeConfig): use runtime config for image repo path
Previously the expected location of the `image_repo` was determined by
an application config setting (which is evaluated at compile time).

The location of the `image_repo` dir is now under the control of the
user. It is expected that the current working dir will contain a
`.devnoderc` file.

The `.devnoderc` file is a YMAL 1.2 document, containing runtime
configuration. Currently the only expected setting in this file is
`image_repo.path`. However it will be utilised for other config settings
in the near future.

The `Devnode.Client.RuntimeConfig` module is responsible for loading and
parsing the config file, which is expected to be in the current working
directory. However in the test environment the path to the runtime
config file is determined by a config setting (in the `devnode_client`
application). In the test environment the runtime config file is an
Embedded Elixir file (.eex) file which is first evaluated before the
YAML is parsed, this allows us to dynamically determine the
image_repo path.