refactor(Client.Scaffold): pass node credentials to build/2
This commit is the first in a series of refactorings targeting the
`Client.Scaffold` module. The aim being to make it more versatile, so
that it can be used when scaffolding the Docker Registry VM.

Previously the `Devnode.Client.Scaffold` module exposed a `build/3`
function that accepted the following arguments:

 * path: project path
 * name: project name
 * image: docker image

This `build/3` function then obtained the credentials for the new node
(using `Devnode.Client.Node.new`).

The `build/3` function has been replaced with a `build/2` function,
which accepts the following arguments:

 * path: project path
 * credentials: full node credentials

Unlike the previous `build/3` function, `build/2` doesn't call
`Devnode.Client.Node.new` itself. Rather it expects the caller to have
made this call. By removing the dependency on `Devnode.Client.Node.new`,
`build/2` is more versatile.

This change also had the benefit of simplifying the `Client.Scaffold`
tests, as there was now no longer a need to mock the
`Devnode.Client.Node.new` function.