refactor(Client.Scaffold): extract ScaffoldMixin
This commit continues a series of refactorings targeting the
`Client.Scaffold` module. The aim being to make it more versatile, so
that it can be used when scaffolding the Docker Registry VM.

A new mixin has been extracted from the `Client.Scaffold` module called
`Client.ScaffoldMixin`. This mixin is to be used by modules that have
scaffolding requirements. Modules that use this mixin, should override
the following functions:

* `tasks/2`: return a list of MFA tuples representing scaffold
operations that can be run in parallel.

* `sub_dirs/0`: return a list names for directories that should be
created.

The original `Client.Scaffold` mixin has been renamed
`Client.NodeScaffold` and the module has been refactored to use the
`Client.ScaffoldMixin`.