refactor(Client.ScaffoldMixin): adopt `Scaffolder` behaviour
Previously the `ScaffoldMixin` defined two functions that were
overridable (`sub_dirs/0` and `tasks/2`). In both cases a noop function
was defined within the `ScaffoldMixin`, meaning that modules that used
the mixin weren't forced to override these functions.

This lack of enforcement made sense with regards the `sub_dirs/0`
function, as scaffold modules may not require subdirectories to be
created.

However it didn't make sense with regards the `tasks/2` function, as
scaffold modules will always require some custom build tasks to be run.

With this in mind, a new behaviour called `Scaffolder` has been defined.
This behaviour enforces modules that are adopting the behaviour to
implement the `tasks/2` function.

The `ScaffoldMixin` adopts the `Scaffolder` behaviour, but doesn't
implement the `tasks/2` function its self. Rather this is left to the
module that uses the `ScaffoldMixin`. If the module in question fails to
implement the `tasks/2` function then a compiler-time waring will be
generated.