feat(Client.Command): Add build-registry
The `Client.CLI` now supports a `build-registry` command. The result of
running this command is to scaffold a VM for running a Docker registry.

The `Client.Command` utilises a new `RegistryScaffold` module for
scaffolding the VM.

The registry VM is configured to use 3 shared directories (`app`, `env`
and `scripts`). Of which the `app` directory is a symlink to the
`image_repo` dir.

The IP address for the registry VM is read from a config variable.