feat(Client.Command): add support for `--force` option to build-registry
By default the `build-registry` command will now output a help message,
if the registry has already been built.

The `build_registry` command now supports a `--force` option (aliased as
`-f`), which will replace an existing registry that has already been
built.

The implementation of this feature involved modifying the `Scaffolder`
behaviour to include an additional function (`valid?/2`), which modules
adopting this behaviour (`RegistryScaffold` and `NodeScaffold`) must
implement.

The `valid?/2` function is called from `build/2` and provides a hook for
scaffolders to conduct validation before the build is commenced. In the
case of the `RegistryScaffold` module `valid?/2` checks if a registry
has already been built.

The `ScaffoldMixin.build/2` function has been updated to now return a
tuple of the form `{:ok, credentials}` if the build completed and
`{:error, fun}` if validation failed. In the case an error tuple is
returned, the `fun` value will be a function that the caller can invoke
to raise an appropriate error.