feat(NodeList): include `port` for vm
The `Devnode.Monitor.NodeList` now expects node entries to contain a
`port` key. This represents the ssh port that port 22 of the vm
should be forwarded to on the host.

The `Devnode.Client.NodeScaffold` module uses this port data when
scaffolding the vagrant file for each node.

The `Devnode.Monitor.NodeList` module uses the `NodePort` module to
determine the next available port and assigns it to the node when it is
added.

The `Devnode.Monitor.NodePort` module looks for an available port
between port 7000 and 65535 (highest available).

The node summary generated by the `Devnode.Client.Command` module now
includes the nodes port information.