refactor(Monitor.NodeServer): use Map.get for reading entry name
Previously the `entry_added/2` function of the `Monitor.NodeServer`
module looked up the `name` field of the new node entry by using
the `new_entry.name` syntax. However Dialyzer complained with the
following warning:

`The call _@1:'name'() requires that _@1 is of type atom() | tuple() not

As `new_entry` is a struct we can use the functions from the `Map`
module. Using `Map.get` to read the `name` field fixed the Dialyzer
warning.