refactor(Monitor.NodeList): change entries type
Previously the `Devnode.Monitor.NodeList` struct's `entries` field
was a `HashDict`. This has been updated to be a `map` instead.

Elixir introduced `HashDict as a workaround for the poor performance
of maps at the time it was written. However, this is fixed now as of
Elixir 1.0.5/Erlang 18.0 and HashDict will be deprecated in future
versions.