rebase(Client.ScaffoldMixin): return standard ok|error tuple from `build/2`
Previoulsy when the `Client.ScaffoldMixin.build/2` returned an error
tuple, it contained an anonymous function (returned from the `valid?/2`
function of the given concrete scaffolder).

`Client.ScaffoldMixin.build/2` has been modified to instead return an
error message. This allows the `Client.Command` module to be the one to
decide when and if an error is raised.

The `Client.Command` module has been updated so that the
`build-registry/1` and `build_node/1` functions raise a standard
`RuntimeError` when the scaffold build fails.