refactor(Client.Command): use result monads for build commands
The `Client.Command.build_node/1` and `Client.Command.build_registry/1`
functions have both been refactored to use a pipeline of functions that
receive and return result monads. This is useful because at each stage
of the pipeline an error can be detected, and subsequent functions will
silently pass. This allows errors to be raised at the end of the
pipeline that are specific to what stage failed.

As a result of this refactoring users receive more specific information
about why their build commands fail.