fix(config): use config/3 and specify application name
Previously `config/2` was being used unconventionally to specify custom
configuration. Whereby an arbitrary key was passed as the `app` argument
and then `Application.get_env` used this key to retrieve the config.

Elixir 1.1 raises a compilation warning when you attempt to configure an
application that is not available (i.e not added as an application
dependency).

The fix is to use `config/3` which provides the correct way to configure
config for a given key for the given application.