fix(dialyzer): remove warnings for string interpolation
Upgrading to Elixir 1.1 with Erlang 18 produced several dialyzer
warnings relating to String interpolation. For example if one of the
variables being yielded was an Atom then dialyzer would complain.

The Elixir core team is intending to fix this issue in Elixir 1.2, but
for the time the fix is to convert the variables in question explicitly
to a String.