fix(Client.Command): check for runtime config before attempt to read
Previously the `build_node/1` function called `ImageRepo.dir` (which
reads from the runtime config), before a check for the existence of the
runtime config was made.

This has been fixed so the `requires_runtime_config/1` function is
called first. This required the `requires_runtime_config/1` to be
refactored to accept an anonymous function, which is only called if the
runtime config exists.