refactor(Monitor.NodeList): represent nodes as a Struct
The `NodeList` module defines a Struct that has an `entries` field which
is a Map containing "nodes". Previously each node was represented as a
simple Map. This has been updated so that each "node" is represented as
a `Monitor.Node` Struct. Representing "nodes" as Structs gives us the
benefit of compile time checks.

Additional Typespecs have been added around this change.