refactor(Client.Command): return result monads rather than raise
Previously the `Client.CLI` ran the `Client.Command.execute/1` function
in a `Task`, trapping the exit signals from this linked process. This
was done because the `execute/1` function was expected to commonly raise
exceptions (as it was designed to communicate command failures in this
way).

This design of "commonly" raising exceptions as a means of communicating
was not a good one. As Jose Valim puts it "Exceptions are for
*exceptional* stuff, things you don't expect". Also by trapping
exits, you are not letting your processes crash.

This has now been refactored so that the `execute` function returns a
result monad which communicates success or failure without the need
for raising exceptions.