feat(Core.Node): add core application to umbrella project
A new library application `devnode_core` has been added to the umbrella
project, this application is intended to contain modules that will be
required by multiple applications within the project as a whole.

The first module to be added to `devnode_core` is the `Core.Node`
module. This module has been extracted from the `devnode_monitor`
application because the intention is to use it in the `devnode_client`
application also (this will be done in a latter commit).