fix(Client.Command): remove struct name from data generated by list cmd
Commit ebb9904c8888ea32c4548a1b93c9382c86dbc4e3 modified
`Monitor.NodeList` so that it represented each node as a struct, rather
than a plan map. This change introduced a bug in the
`Client.Command.list_nodes/1` function, namely that the list output
would include the hidden `__struct__` field value for each node struct.

The test that should have picked this bug up in `CLI.Test` was missing
an `assert` and hence the issue went unnoticed. The test has been
updated and now relies upon the `Core.Node` module.