fix(Client.CLI): prevent Task.await timeouts
The `Devnode.Client.CLI.build_response/1` function runs the
`Devnode.Client.Command.execute/1` function in a separate process using
`Task.async/1`. `Task.await/2` is used to await for the task reply. The
default timeout for this await is 5000 milliseconds. This timeout has
been updated to be infinite, to accommodate commands that prompt the
user for input.