refactor(Client.Command): rename and extract routes to seperate modules
The `Client.Command` module has been renamed to `Devnode.Router`. This
module is solely responsible for inspecting the options passed to the
devnode CLI and determining the appropriate route to invoke.

The different routes are each represented by a separate module that
exposes an `execute/1` function.

The `result_monad` typespec has been extracted into the `Core.Types`
module, as this type needs to be used across multiple modules (and
potentially multiple applications). The ideal home for this typespec
would be the `Towel` module, however this module is a third party
dependency.