feat(Client.BuildRoute): validate that image repo exists
The `BuildRoute` route now validates for the existence of an image repo.
If an image repo doesn't exist then help information is output,
informing the user of this requirement.

This feature required a bit of refactoring of the `BuildRoute` module,
to support multiple validations; as this route also validates for the
presence of a runtime config file. To accomplish this a new module
called `RouteValidation` is used.

The `RouteValidation.validate/1` function takes a list of predicate
functions. If validation is deemed to be successful then each of the
predicate functions must evaluate to `true`. If validation fails then an
error result monad is returned, containing the name of the first
validation function that failed. `Client.CLI` uses this information to
select the appropriate help information to display.