feat(Client.GenerateRuntimeConfigRoute): add route
A new route `GenerateRuntimeConfigRoute` has been added that is
triggered using the following command:

```bash
devnode_client generate runtime-config
```

This new route will generate a `devnoderc` file in the current/working
directory. The purpose of a `devnoderc` file is to provide runtime
configuration for devnode.

This new route validates the existence of an image repo, and will return
help content if one doesn't exist.